requirements.txt
amqp==2.5.2
apns2==0.5.0
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
billiard==3.6.1.0
black==19.10b0
cachetools==3.1.1
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.0
Django==3.0
django-environ==0.4.5
git+https://github.com/htrueman/django-push-notifications.git
djangorestframework==3.11.0
drf-yasg==1.17.0
google-api-python-client==1.7.11
google-auth==1.9.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httplib2==0.14.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.3.0
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==19.2
pathspec==0.6.0
psycopg2-binary==2.8.4
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
pytz==2019.3
pywebpush==1.10.1
regex==2019.12.9
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
whitenoise==5.0.1
zipp==0.6.0
zope.interface==4.7.1
